re-render training book
Re-render of training/_book with the updateApril2026 chapter rewrites:
pak::pak bootstrap + setupProject pattern across all chapters, plus the
new Experiments / ExperimentsAndScenarios / FireUnderClimateChange Part.
